Netlify docs (#70)
* catch errors in load dataframe

* avoid temp dir to be deleted

* updated docs for github pages

+++ tests/test_images_io.py
@@ -20,20 +20,21 @@ def setup_module(module):
 def setup_function(func):
     ''' Setup for test functions '''
     for filename in os.listdir(cache_directory):
-        file_path = os.path.join(cache_directory, filename)
-        try:
-            if os.path.isfile(file_path) or os.path.islink(file_path):
-                os.unlink(file_path)
-            elif os.path.isdir(file_path):
-                shutil.rmtree(file_path)
-        except Exception as e:
-            print('Failed to delete %s. Reason: %s' % (file_path, e))
+        if (filename != '__emptyfile__'):     
+            file_path = os.path.join(cache_directory, filename)
+            try:
+                if os.path.isfile(file_path) or os.path.islink(file_path):
+                    os.unlink(file_path)
+                #elif os.path.isdir(file_path):
+                #    shutil.rmtree(file_path)
+            except Exception as e:
+                print('Failed to delete %s. Reason: %s' % (file_path, e))
     # cache_directory = os.path.join(os.getcwd(), 'tests/resources/temp/io')
     #if os.path.exists(cache_directory):
     #    shutil.rmtree(cache_directory)
     #os.mkdir(cache_directory)
     #os.chdir(cache_directory)
-
+    #os.mkdir(cache_directory)
 
 def test_load_single_image():
     image_file_name = 'tests/resources/images/lungs/CHNCXR_0001_0.png'
@@ -94,6 +95,12 @@ def test_load_images_extensions():
 
 def test_image_url_to_memory():
     image = ami.load_image_from_url(image_url)
+    assert len(image.shape) == 3
+
+
+def test_image_url_resize_memory():
+    img_path = 'https://charting.vwdservices.com/tchart/tchart.aspx?user=Tijdnet&issue=360017012&layout=gradient-v1&startdate=3Y&enddate=today&res=endofday&width=876&height=400&format=image/png&culture=nl-BE'
+    image = ami.load_image_from_url(img_path)
     assert len(image.shape) == 3
 
 def test_image_url_to_cache():
